Refactor pre tools, smooth out build process
